[Update] favorites,likes非同期通信化

+++ app/controllers/employee/favorites_controller.rb
@@ -2,17 +2,13 @@ class Employee::FavoritesController < ApplicationController
 
   def create
     @post = Post.find(params[:post_id])
-    favorite = @post.favorites.new(employee_id: current_employee.id)
-    favorite.save
-    flash[:notice] = "いいねしました"
-    redirect_to request.referer
+    current_employee.favorites.find_or_create_by(post_id: @post.id)
   end
 
   def destroy
     @post = Post.find(params[:post_id])
     favorite = current_employee.favorites.find_by(post_id: @post.id)
-    favorite.destroy
-    redirect_to request.referer
+    favorite.destroy if favorite
   end
 
 end
